+++ javaScript/projects/binary-searh-trees/src/containers/node.js
@@ -1,0 +1,7 @@
+export default class Node {
+  constructor(data) {
+    this.leftNode = null;
+    this.rightNode = null;
+    this.data = !data ? null : data;
+  }
+}